fix: show new Select in navigation
